cosmic-de/cosmic-wallpapers: Use the `mirror` option for git and LFS clones
Due to the way things are setup at the moment, we reuse the same
`DISTDIR` repo for multiple versions (e.g.: alpha2 -> alpha3 -> alpha4).

Problem is that every time we emerge a new version, the repo with
`CLONE_TYPE=single` does not contain all other refs and is not reset, so
the user needs to manually remove it and reclone it.

While `CLONE_TYPE=mirror` increases repo size, for cosmic-wallpaper it
takes ~18MB. It's still better for the user to cleanup once/year (or
whenever) instead of every time a new version hits the repo. Old/deleted
refs are purged in this mode.

Fixes #45